improved left/right keys non/reflection

+++ includes/rt.h
@@ -6,7 +6,7 @@
 /*   By: eLopez <elopez@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/21 00:06:40 by eLopez            #+#    #+#             */
-/*   Updated: 2018/01/16 21:27:13 by elopez           ###   ########.fr       */
+/*   Updated: 2018/01/16 22:48:11 by elopez           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -161,7 +161,7 @@ typedef struct	s_objects
 	int					type;
 	t_union				u;
 	t_vect				norm;
-	double				shine;
+	int					shine;
 	double				spec;
 	double				diff;
 	double				m;
